Fix signed/unsigned warning.

+++ asio/include/asio/impl/config.hpp
@@ -39,7 +39,9 @@ T config_get(const config_service& service, const char* section,
       char* end = nullptr;
       errno = 0;
       unsigned long long result = std::strtoull(str, &end, 0);
-      if (errno == ERANGE || result > (std::numeric_limits<T>::max)())
+      if (errno == ERANGE
+          || result > static_cast<unsigned long long>(
+            (std::numeric_limits<T>::max)()))
         detail::throw_exception(std::out_of_range("config out of range"));
       return static_cast<T>(result);
     }
